p2p: Redesign handling of timeouts in ChainSyncer
We were seeing timeouts too frequently and because of that I'd
introduced an eager retry mechanism, but it turns out those timeouts
were because we had too many cpu-intensive tasks that were blocking the
main thread and thus preventing the asyncio main loop from processing
received msgs fast enough. Now that those cpu-intensive tasks are
performed in sub-processes, we rarely see timeouts and in those cases it
generally means the peer has disconnected so instead of retrying we
simply abort the sync with that peer, which causes it to be resumed with
a different peer.