p2p: PeerPool no longer restarts itself on unexpected exceptions
PeerPool's main loop used to catch any exceptions, log them, restart all
peers and continue in the loop. This was useful because our services
generally run the PeerPool in the background (with `asyncio.ensure_future`),
but it turns out this is problematic because sometimes when exiting the
main process the PeerPool's main loop may get an exception
(a RuntimeError, because the event loop has been closed) before the
parent service has had a chance to tell the PeerPool to shut down, so we
end up in an infinite loop with the PeerPool catching those
RuntimeErrors and trying to restart all peers.